Reorganize code into modules and fix __dirname incompatibility
