+++ src/resolve-source-to-id.js
@@ -1,0 +1,7 @@
+export const resolveSourceToId =
+  ({ importTag }) =>
+  (source) => {
+    if (source !== importTag) return null;
+    // If source is tagged, return it so the import plugin can run against it
+    return source;
+  };
