+++ test/fixtures/parse-fixture-data.js
@@ -1,0 +1,5 @@
+import { readFileSync } from "fs";
+import { resolve } from "path";
+
+export const parseFixtureData = (filename) =>
+  readFileSync(resolve(process.cwd(), "test/fixtures", filename), "utf8");
